Pool Completion keys for V2 Add and Read
Based on changes on the Yahoo branch, updated to fit with PCBC refactor.
https://github.com/yahoo/bookkeeper/commit/3cb38fdae6bbba17bd4ea590516a761737896275

A recycler is used to pool V2CompleteKeys. Once the key has been
acquired, its ownership is passed to the corresponding
CompletionValue. When the completion value callback runs, which is a
signal that the completion is done, the key is released back into the
pool.