BOOKKEEPER-883: Test timeout in bookkeeper-benchmark
Problem:

The BenchReadThroughputLatency is tight with FlatLedgerManager. so lots of assumptions are made based on how the znodes are changed when ledgers are created. There was a change introduced LedgerIdGenerator, which broke the assumptions that made by BenchReadThroughputLatency.

Fix:

- Use a hashset to cache processed ledgers on reacting on children changes
- Remove unpredictable test on next ledger
- Fix an error logging on FlatLedgerManager processing ledgers

Author: Sijie Guo <sijie@apache.org>

Reviewers: Matteo Merli <mmerli@apache.org>

Closes #10 from sijie/BOOKKEEPER-883

(cherry picked from commit 8729d12be50295086e7440dfa5d0256abb7688d5)
Signed-off-by: Sijie Guo <sijie@apache.org>