bookkeeper start script needs to be portable.
The script was using `ps -p` to test if a process is running. The `-p`
option is not available on all implementations of `ps` (e.g.:
busybox). Instead, use `kill -0`, which is more portable to verify if a
PID is running.

Replace backticks with subshell, which is also more portable.