BOOKKEEPER-937: Upgrade protobuf to 2.6
protobuf/protoc 2.4 cannot be installed with brew on mac and building it on mac always result is build errors hence leaves an option of switching to linux to run protoc.
I upgraded to 2.6 instead. It is compatible with 2.4 on the wire and shaded so should not create any problems. All tests passed.
Please ignore changes in java files in attached patch during review; these are auto-generated.

protobuf 2.4 in compat* dependencies is shaded now to not conflict with 2.6.

Author: Andrey Yegorov <ayegorov@salesforce.com>

Reviewers: eolivelli@gmail.com <eolivelli@gmail.com>, zhaijia03@gmail.com <zhaijia03@gmail.com>, sijie@apache.org <sijie@apache.org>

Closes #51 from dlg99/task/BOOKKEEPER-937