BOOKKEEPER-965: Long Polling Part I: Changes in the write path
This is the first in the series of changes for enabling long polling between bookkeeper client and the bookkeeper server. The changes were originally implemented in the Twitter fork and these pull request combine multiple commits from Twitter's bookkeeper fork as they include not only the changes made initially but also bug fixes added since.

The first change captures the changes on the write path (AddEntry). We track the last add confirmed in the FileInfo so that we can trigger actions when the value is updated

Author: Robin Dhamankar <robindh@Robins-MacBook-Air.local>

Reviewers: sijie@apache.org <sijie@apache.org>

Closes #73 from robindh/LongPollWritePath