BOOKKEEPER-1004: Allow bookie garbage collection to be triggered manually from tests
Originally done by Ivan Kelly

The current gc tests rely on waiting on a timeout for gc to run. It's
never certain whether it has run or not or if it's still running.
This patch allows tests to trigger a gc run and gives the client
a future to know when it has completed. I haven't changed the gc
algorithm, but now it runs in a scheduled executor rather than as a
Thread.

BOOKKEEPER-1004: Allow bookie garbage collection to be triggered manually from tests

Originally done by Ivan Kelly

The current gc tests rely on waiting on a timeout for gc to run. It's
never certain whether it has run or not or if it's still running.
This patch allows tests to trigger a gc run and gives the client
a future to know when it has completed. I haven't changed the gc
algorithm, but now it runs in a scheduled executor rather than as a
Thread.