BOOKKEEPER-1042: Deploy SNAPSHOTS of master to public Apache Repository
Make all pom.xml files have a -SNAPSHOT version.

Currently we have the 'compats' subprojects which have a strange versioning scheme.
Actual versioning scheme for that poms will clash will other versions of the project, because they are fixed and do not depend on the common parent pom.

Changing the version will make the build more reproducible and it is a prerequisite for the deployment of -SNAPSHOT version to the official Apache Snapshots Repository

Author: Enrico Olivelli <eolivelli@apache.org>

Reviewers: Matteo Merli <mmerli@apache.org>, Sijie Guo <sijie@apache.org>

Closes #134 from eolivelli/BOOKKEEPER-1042-pom-for-snapshots