BOOKKEEPER-1051: Fast shutdown for GarbageCollectorThread
Several unit tests are taking very long time to complete (eg: `BookieLedgerIndexTest` taking ~10 minutes).
The reason is that these tests are playing with the ZK quorum shutting it down and after the test succeeds, the shutdown phase is taking long time, since we try to do graceful shutdown with 1min wait time.
I think is better to interrupt immediately the garbage collector thread when shutting down the bookie.

Author: Matteo Merli <mmerli@apache.org>

Reviewers: Enrico Olivelli <None>

Closes #141 from merlimat/fast-gc-thread-shutdown