BOOKKEEPER-1055: Optimize handling of masterKey in case it is empty
On each request client and bookies are exchanging the ledger masterKey, which is a 20 bytes MAC digest of the ledger password.

For each request there is a considerable overhead in allocating byte arrays when parsing the add/read requests.

If the client is a passing an empty password, we should optimize the data path to skip all allocations (related to the masterKey) and instead rely on a static byte array.

Author: Matteo Merli <mmerli@apache.org>
Author: Matteo Merli <mmerli@yahoo-inc.com>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Venkateswararao Jujjuri (JV) <None>

Closes #156 from merlimat/empty-password