BOOKKEEPER-1090: Use LOG.isDebugEnabled() to avoid unexpected allocations
Using `LOG.debug(...)` can lead to multiple unexpected memory allocation, even when the logger it's turned off.
For example, int and long parameter are boxed into Integer and Long objects and the var-arg parameters are using an `Object[]` to hold
them.
We should guard all usages of `LOG.debug()` with the `if (LOG.isDebugEnabled()` guard.

Author: Matteo Merli <mmerli@yahoo-inc.com>

Reviewers: Enrico Olivelli, Sijie Guo

Closes #179 from merlimat/if-debug