Add Http Server for Bookkeeper.
Provide a general interface for HttpServer, which can be easily implemented by different HTTP frameworks.
This change include two implementations of HttpServer, one is TwitterServer and another one is Vertx.

Provide a general AbstractHandlerFactory, which is able to create handlers to handler different http apis,
and is able to be implemented for different HTTP frameworks.

Provide a service level classes, which include all the logics that http handlers will use.

Add test case for HttpServer