ISSUE #263: Enable request rate limiter in zookeeper client
Descriptions of the changes in this PR:

Eanble request rate limiter in zookeeper client. The rate limiter logic is in place, but is never enabled. This change is to expose the settings in configurations and enable them in zookeeper client at both client and server.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <None>, Jia Zhai <None>

This closes #264 from sijie/interface_syncs, closes #263