BOOKKEEPER-1104: Fix testWithDiskFullAndAbilityToCreateNewIndexFile
The usage threshhold was chosen to be very close to the actual disk
usage at test time.  This made the test unnecessarily fragile in the
case that there other things concurrently using the disk.  Since we
aren't really testing that here, simply set the threshhold to be really
low.

Signed-off-by: Samuel Just <sjustsalesforce.com>

Author: Samuel Just <sjust@salesforce.com>

Reviewers: Jia Zhai <None>, Matteo Merli <None>

This closes #310 from athanatos/forupstream/BOOKKEEPER-1104