ISSUE #390: [CI] Test errors in TestRackawareEnsemblePlacementPolicyUsingScript
Descriptions of the changes in this PR:

`Shell` class was ported from hadoop, but it contains a lot of unused code. Especially there is not `HADOOP_HOME` will set, so an exception will be thrown. (although the issue seems to be hidden in mac)

This change is to clean up the `Shell` class. The `Shell` class is only used for executing shell script configured in rackaware placement policy.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Jia Zhai <None>, Enrico Olivelli <None>

This closes #396 from sijie/sijie/cleanup_shell, closes #390