ISSUE #422: Don't use community layout for documentation pages
Descriptions of the changes in this PR:

The layout of documentation page is marked as community. when users navigate to this page, it is a bit difficult for the users to figure out where they are. We should stick documentation pages to doc layout.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Jia Zhai <None>

This closes #423 from sijie/documentation_index, closes #422 and squashes the following commits:

4ec16d6 [Sijie Guo] remove community layout for docs page
21320b4 [Sijie Guo] The index page for documentation should use documentation layout rather than community layout