ISSUE #304: New Release Guide
Descriptions of the changes in this PR:

This is an attempt to improve the current release guide base on a few references :  [Beam](https://beam.apache.org/contribute/release-guide/) [Pulsar](https://github.com/apache/incubator-pulsar/wiki/Release-process)

Add this in main repo for reviews. We can later on make it part of website once #293 is completed.

A few items are TBD. comments are welcome.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>

This closes #305 from sijie/release_guide, closes #304