ISSUE #462: TestTLS failures
Descriptions of the changes in this PR:

Problem:

The behavior of loading configuration in `AbstractConfiguration` was changed recently.
So the TLS provider was not cleared correctly for the failed tests.

Solution:

Explicitly clear TLS provider for each configuratio to address the failed tests.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>

This closes #478 from sijie/fix_tls_test, closes #462