DL-45: DL should allow ByteBuf based API to avoid copying bytes array
Descriptions of the changes in this PR:

This change leverages the `ByteBuf` api introduced in bookkeeper 4.5.0. It will avoid copying bytes array between LogRecord and LogRecordSet/Entry, and avoid copying bytes from DL to BK client.

This change also bump the lz4 library to `1.3.0` to leverage the `ByteBuffer` binding.

Author: Sijie Guo <sijie@apache.org>
Author: Jia Zhai <zhaijia03@gmail.com>
Author: Jia Zhai <zhaijia@apache.org>

Reviewers: Jia Zhai <None>, Enrico Olivelli <eolivelli@gmail.com>, Leigh Stewart <None>

This closes #151 from sijie/zero_copy