BOOKKEEPER-1106: Introduce write FileInfo cache and read FileInfo cache
Problem: when read behind happens, it would quickly read bunch of ledgers, which will evict current active ledgers for writing from the ledger cache. with the ledger being evicted from cache, it would impact the write performance.

This feature is contributed by sijie , in which we introduced write file info cache and read file info cache to cache the ledger index separately for read and write, so that when catch up read happens, it will not evict the file info for writes.

Author: Yiming Zang <yzang@twitter.com>

Reviewers: Enrico Olivelli <eolivelli@apache.org>, Jia Zhai

This closes #513 from yzang/yzang/BOOKKEEPER-1106