ISSUE #527: Introduce backoff and retry utilities
Descriptions of the changes in this PR:
Mainly want to  Introduce backoff and retires utilities:
Backoff implements various backoff strategiesis, and id intended to determine the duration after which a task is to be retried.
Retries is a util class for supporting retries with customized backoff.

Author: Jia Zhai <zhaijia@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Sijie Guo <sijie@apache.org>

This closes #528 from zhaijack/retries_backoff, closes #527