ISSUE #197: Provide a guide for running on k8s
Descriptions of the changes in this PR:

add one page for how to deploy on k8s. it is based on http://bookkeeper.apache.org/docs/latest/deployment/kubernetes/ and use distributedlog image and add instructions on how to create distributedlog namespaces and run benchmark.

This change is based on #196

Author: Sijie Guo <sijie@apache.org>

Reviewers: Jia Zhai <None>

This closes #198 from sijie/add_docker, closes #197