BOOKKEEPER-1062: Fixed the race in AuditorLedgerCheckerTest
Descriptions of the changes in this PR:

This fixes the race condition in the auditor between read only bookies and available bookies.  But because I am not sure of the performance impacts to modifying getReadOnlyBookies, I have opted to add in a new getROBookies.  But I am happy to change it.  This pull request is mostly to show my work and see what others think.

Author: Robert Evans <evans@yahoo-inc.com>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>, Sijie Guo <sijie@apache.org>

This closes #556 from revans2/BOOKKEEPER-1062