ISSUE #226: ByteBuf.release() was not called before it's garbage-collected
Descriptions of the changes in this PR:

the problem is a new entry buffer was allocated when closing log segment writer. the entry buffer is never used and also never recycled. It causes an annoying logging.

the fix is to assign a dummy entry writer which basically rejects writes, when closing a log segment writer. it would prevent leaking bytebuf.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Jia Zhai <None>

This closes #230 from sijie/fix_bytebuf_release_pr, closes #226