Fail earlier when encountering DNS issues
We had some DNS issues today and various TestRegionAwareEnsemblePlacementPolicy tests were failing for me.

The LOG warning in RackawareEnsemblePlacementPolicyIml was not printing the right variable, which would have given a clue.

Additionally, the repp.myRegion was set to UNKNOWN, causing incorrect code flow.  Added an assert to kill the test as soon as this occurs.

Author: Aaron Gresch <agresch@yahoo-inc.com>

Reviewers: Ivan Kelly <ivan@ivankelly.net>, Enrico Olivelli <eolivelli@gmail.com>, Sijie Guo <sijie@apache.org>

This closes #658 from agresch/agresch_dns_failure