DistributionSchedule uses int[] rather than HashSet
This avoids the autoboxing on Integers and allocations of cells for
the hashset.

This also implies using int[] for the write set in the
DistributionSchedules and PlacementPolicies.

This patch was originally submitted as dc7933b1 on the yahoo-4.3
branch, though this has been modified extensively to remove the
dependency on carrotsearch hppc and to allow it to work with placement
policies which didn't exist at the time of the original patch.