ISSUE #666: Introduce registration client for bookkeeper client to do bookie discover
Descriptions of the changes in this PR:
while in BOOKKEEPER-628, It is trying to improve/generalize the bookie registration process.

This PR follows the work of #663, which provide bookie side registration, and it is for client side work, it tries to introduce RegistrationClient for bookkeeper client to do bookie discover.

This PR follows the work of #663, which provide bookie side registration,

Author: Jia Zhai <zhaijia@apache.org>

Reviewers: Sijie Guo <sijie@apache.org>

This closes #667 from zhaijack/client_registration2, closes #666