ISSUE #726: Unit Tests failure with BKMetadataVersionException
Descriptions of the changes in this PR:

Problem:

The problem is introduced after making storing ctime optional. the verification become inconsistent because the way how ctime was assigned and compared. It causes the test cases failing in #726

Solution:

This change is to change `isConflict` to ignore comparing ctime if client disables storing system time as ctime.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>

This closes #730 from sijie/fix_ctime_issue, closes #726