ISSUE #734: bookkeeper-benchmark is broken after protobuf files are moved to bookkeeper-proto
Descriptions of the changes in this PR:

Problem:

The issue was introduced after protobuf files are moved to bookkeeper-proto. so the protobuf is not shaded correctly.
So in bookkeeper-benchmark, it is conflicted with the protobuf version introduced by hadoop.

Solution:

- update the bookkeeper-server pom.xml to shade protobuf correctly
- remove hadoop dependency from bookkeeper-benchmark.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Ivan Kelly <ivan@ivankelly.net>, Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>

This closes #735 from sijie/protobuf_issue, closes #734

(cherry picked from commit 9b538aff7da5cba15ae59b70e0b0c9bd2464dca3)
Signed-off-by: Jia Zhai <zhaijia@apache.org>