BookieNettyServer: add contextHandler to the local transport pipeline
Left off of 8e0bd2c3d81b522e97434d8646915f36422a104b.  In fact,
authentication is already enabled on LocalTransport.  This extra line is
needed for the machinery which allows auth plugins to access ssl state.
Currently, the only plugin which uses that machinery is in TestTLS.
Adds test cases to validate that functionality with Local Transport
enabled.

Signed-off-by: Samuel Just <sjust@salesforce.com>