ISSUE #739: Move the implementation of http services to be under server package
Descriptions of the changes in this PR:

bookkeeper-http module provides the framework for http admin rest api. it is using 'org.apache.bookkeeper.http' and 'org.apache.bookkeeper.http.service. the http package inbookkeeper-server` is reusing these two packages. it is not good for two modules sharing same package name, it is going to be confusing javadoc.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>

This closes #740 from sijie/http_impl, closes #739