ISSUE #693: add interface and implementation of LedgerEntries
In the new API proposed in #506 , we return Iterable<LedgerEntry> for read entries. It is a bit problematic when returning multiple iterators and have no ability to release the entry buffers hold by ledger entries.
In this change, the implementation of retainIterator() increased the reference of the buffers when creating an iterator, it will be easy for byteBuf's management; it also make LedgerEntries a recycle object, and it is returned to the pool when #close() is called. on #close(), it also releases all the references to release resources.

Descriptions of the changes in this PR:
1. add interface LedgerEntries;
2. add implemtation of LedgerEntries;
3. fix related build and test errors.

Author: Jia Zhai <zhaijia@apache.org>

Reviewers: ivankelly, sijie, eolivelli

This closes #727 from zhaijack/issue-693, closes #693