Tests wait for client to see started bookie
Up until now we were only waiting for the bookie to exist in the
bookie list, after which we triggered a read. However, when the
registration changes went in, the read wasn't actually populating the
mechanism used to select bookies, so we ended up momentarily having
unusable bookkeeper clients until the correct read triggered. This
resulted in flakiness in random tests.

This change makes it so we wait until the client has seen a bookie
before startBookie will return. It also makes sure that reading the
bookie list actually populates the bookie selection mechanism.