Make distributedlog compiled with latest bookkeeper version
Descriptions of the changes in this PR:

There are code changes on PendingReadOp for new bookkeeper api in current master. DistributedLog uses PendingReadOp for some administration tools. So the current master doesn't compile with the latest bookkeeper version. This code change is to fix that.

The change here includes:

- bump bk to 4.7.0-SNAPSHOT (will switch to 4.6.0 after it is released)
- change to use the latest CompletableFuture in latest PendingReadOp. (this change doesn't target at making distributedlog work with new API)

Author: Sijie Guo <sijie@apache.org>

Reviewers: Jia Zhai <None>

This closes #240 from sijie/sijie/use_new_ledger_api