ISSUE #230: Enable Checkstyle on the proto package
This is part of #230 and adds Checkstyle verification to the proto
package in bookeeper-server.

The changes here should generally be non-controversial, but I would like to highlight one change for particular review. In the `PerChannelBookieClient` class, there is a `public static final` field called `txnIdGenerator`. The checkstyle rules require that such public fields are ALL_CAPS. So one option would be to change that field to `TX_ID_GENERATOR` (this would be a change to the public API).

The relevant diff is here: https://github.com/apache/bookkeeper/compare/master...acoburn:checkstyle_proto?expand=1#diff-e50ee2c1aec1539ea185a94605b0e550R143

However, that field is never used outside this class, so the other option is to make the field private and keep the name as-is. (Making the field private is also a change to the public API)

I chose the second option because the field seemed more like an internal field. If, however, the field should be left public, I can add a commit that changes the name to `TX_ID_GENERATOR` so that it passes the checkstyle rules.

Author: Aaron Coburn <acoburn@amherst.edu>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>, Sijie Guo <sijie@apache.org>

This closes #792 from acoburn/checkstyle_proto, closes #230