ISSUE #230: Enable checkstyle for more packages
Part of #230, this enables the checkstyle plugin for the `replication` and
`server.http.service` packages.

Descriptions of the changes in this PR:

Most of the changes here are merely cosmetic. The only change of any significance is in the `RecoveryBookieService.RecoveryRequestJsonBody` class in which the JSON object fields had to be changed in order to conform to the Checkstyle rules: `bookie_src` to `bookieSrc` and `delete_cookie` to `deleteCookie`. Using a Jackson annotation (`JsonProperty`), however, means that these field names continue to be parsed and serialized as `bookie_src` and `delete_cookie`, respectively.

Author: Jia Zhai <jiazhai@users.noreply.github.com>
Author: Aaron Coburn <acoburn@amherst.edu>

Reviewers: Jia Zhai <None>, Sijie Guo <sijie@apache.org>

This closes #800 from acoburn/checkstyle_replication, closes #230