ISSUE #230: Enable checkstyle on some test classes
Part of #230, this enables checkstyle verification on some
of the test code. This does not do checkstyle validation on
the client or bookie package test code.

Enabling checkstyle on the entire test codebase would result in a _very_ large PR. This is an attempt to break it up into smaller pieces. The changes here are almost exclusively related to whitespace, javadocs and import order.

One public change of note is in `replication.AuditorLedgerCheckerTest`: the `_testDelayedAuditOfLostBookies` method is changed to `testInnerDelayedAuditOfLostBookies`.

Author: Aaron Coburn <acoburn@amherst.edu>

Reviewers: Jia Zhai <None>, Sijie Guo <sijie@apache.org>

This closes #819 from acoburn/checkstyle_test, closes #230