Use netty DefaultThreadFactory everywhere to take advantage of fast thread local access
Merging https://github.com/yahoo/bookkeeper/commit/38ff21ad8bbd8ac8ec8a4898ef9ecf473fea94b7 from Yahoo branch.

Netty has a special `FastThreadLocal` class that is used in many places (eg: when borrowing a buffer from a pool). The mechanism in which it works is that the thread has to be created from a particular thread factory to take advantage of it (it adds a couple of fields subclassing the `Thread` class).

Author: Matteo Merli <mmerli@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Sijie Guo <sijie@apache.org>

This closes #823 from merlimat/thread-factory