Client times out requests in batch rather than individually
This patch is a squash of f59b597f, 00fc5cd2, 1658eb61 & 833dc307 from the yahoo-4.3 branch. The change removes the HashWheelTimer from PerChannelBookieClient and instead runs through all outstanding ops periodically, timing them out as necessary.

The motivation for this change is to reduce the number of objects created per add/read request. With HashWheelTimer a TimerTask needs to be added to the Timer, which allocates and returns a Timeout object. We had to be careful to cancel this timeout on success.

The new approach is simpler, doesn't require any allocation per request, and doesn't require any cancellation. Functionally the end result is the same - very slow requests are timed out.

The patch also makes some additions to the original changes (due to the codebase moving since it was originally written).
- PendingAddOps are changed to use the same mechanism for quorum timeouts.
- The checkTimeout method is moved to the PCBCPool so we can run the check on non-connected PCBC instances. This is necessary for TLS.

Author: Ivan Kelly <ivank@apache.org>
Author: Siddharth Boobna <sboobna@yahoo-inc.com>
Author: Matteo Merli <mmerli@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>

This closes #817 from ivankelly/yahoo-bp-9